MINI PRJOECT
This Zip file contains (in csv format) (Link1 or Link2 or Link3) names given to babies, both male and female and their rankings of a particular country for every year from 1944 to 2013. Create an interface which can display the following queries:

Popular Names by Birth Year: User can enter a year and choice like Top 10, Top 20..Top 1000 etc and the relevant names will be displayed. Choices should be there for only males, only females and both. Name rankings should also include Number of Births
Popularity of a name that has changed over the years: Given a name and a year it should show (preferably in a graphical format) how popularity has changed starting from that particular year till 2013. For example if a user inputs: Brendon and a year 2000 it should show the popularity of the name starting from 2000 to 2013. Again choice of Male and Female should be there because some names can be both male and female.

user is supposed to enter:-
   the year, gender and no of names and click 'submit', to fetch the popular names.
   the year, gender and name and then click 'graph', to obtain the graph showing popularity of the name from the year entered to 2013.